Move some methods to region_infer/mod.rs
